initial folder refactor
