+++ ui/app/src/main/java/com/example/wat2take/MainActivity.kt
@@ -1,32 +1,20 @@
 package com.example.wat2take
 
-import android.graphics.Color
 import android.os.Bundle
-import android.util.Log
-import androidx.compose.material.*
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Icon
 import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.key.Key.Companion.Home
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavController
-import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import com.example.wat2take.viewmodels.TranscriptDataStore
 import com.example.wat2take.navigation.BottomNav
-import com.example.wat2take.navigation.BottomNavItem
 import com.example.wat2take.ui.theme.Wat2TakeTheme
 import com.example.wat2take.welcome.WelcomeCourses
 import com.example.wat2take.welcome.WelcomePage
