+++ ui/app/src/main/java/com/example/wat2take/UploadTranscript.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.example.wat2take.Global.Companion.SERVER_URL
+import com.example.wat2take.viewmodels.TranscriptDataStore
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.rememberMultiplePermissionsState
 import com.google.gson.JsonParser
