+++ ui/app/src/main/java/com/example/wat2take/welcome/WelcomeCourses.kt
@@ -1,11 +1,8 @@
 package com.example.wat2take.welcome
 
-import android.annotation.SuppressLint
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -17,9 +14,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
-import com.example.wat2take.Global
 import com.example.wat2take.R
-import com.example.wat2take.TranscriptDataStore
+import com.example.wat2take.viewmodels.TranscriptDataStore
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
